Colocate waterfalltaskstatusicon files
